show names of the trempists while click on marker on map
